review refactoring
