+++ lib/collection/src/collection_manager/holders/proxy_segment.rs
@@ -15,8 +15,10 @@ use segment::data_types::facets::{FacetParams, FacetValue};
 use segment::data_types::named_vectors::NamedVectors;
 use segment::data_types::order_by::OrderValue;
 use segment::data_types::query_context::{QueryContext, SegmentQueryContext};
+use segment::data_types::segment_manifest::SegmentManifest;
 use segment::data_types::vectors::{QueryVector, VectorInternal};
-use segment::entry::entry_point::{SegmentEntry, SegmentManifest};
+use segment::entry::entry_point::SegmentEntry;
+use segment::entry::partial_snapshot_entry::PartialSnapshotEntry;
 use segment::index::field_index::{CardinalityEstimation, FieldIndex};
 use segment::json_path::JsonPath;
 use segment::telemetry::SegmentTelemetry;
@@ -346,6 +348,21 @@ impl ProxySegment {
         )?;
 
         Ok(result.into_iter().next().unwrap())
+    }
+}
+
+impl PartialSnapshotEntry for ProxySegment {
+    fn take_partial_snapshot(
+        &self,
+        _temp_path: &Path,
+        _tar: &tar_ext::BuilderExt,
+        _manifest: &SegmentManifest,
+    ) -> OperationResult<()> {
+        todo!()
+    }
+
+    fn get_segment_manifest(&self) -> OperationResult<SegmentManifest> {
+        todo!()
     }
 }
 
@@ -1266,19 +1283,6 @@ impl SegmentEntry for ProxySegment {
         )?;
 
         Ok(())
-    }
-
-    fn take_partial_snapshot(
-        &self,
-        _temp_path: &Path,
-        _tar: &tar_ext::BuilderExt,
-        _manifest: &SegmentManifest,
-    ) -> OperationResult<()> {
-        todo!()
-    }
-
-    fn get_segment_manifest(&self) -> OperationResult<SegmentManifest> {
-        todo!()
     }
 
     fn get_telemetry_data(&self, detail: TelemetryDetail) -> SegmentTelemetry {
