+++ lib/segment/src/entry/entry_point.rs
@@ -12,6 +12,7 @@ use crate::data_types::named_vectors::NamedVectors;
 use crate::data_types::order_by::{OrderBy, OrderValue};
 use crate::data_types::query_context::{QueryContext, SegmentQueryContext};
 use crate::data_types::vectors::{QueryVector, VectorInternal};
+use crate::entry::partial_snapshot_entry::PartialSnapshotEntry;
 use crate::index::field_index::{CardinalityEstimation, FieldIndex};
 use crate::json_path::JsonPath;
 use crate::telemetry::SegmentTelemetry;
@@ -25,7 +26,7 @@ use crate::types::{
 ///
 /// Assume all operations are idempotent - which means that no matter how many times an operation
 /// is executed - the storage state will be the same.
-pub trait SegmentEntry {
+pub trait SegmentEntry: PartialSnapshotEntry {
     /// Get current update version of the segment
     fn version(&self) -> SeqNumberType;
 
@@ -316,74 +317,9 @@ pub trait SegmentEntry {
         snapshotted_segments: &mut HashSet<String>,
     ) -> OperationResult<()>;
 
-    fn take_partial_snapshot(
-        &self,
-        temp_path: &Path,
-        tar: &tar_ext::BuilderExt,
-        manifest: &SegmentManifest,
-    ) -> OperationResult<()>;
-
-    fn get_segment_manifest(&self) -> OperationResult<SegmentManifest>;
-
     // Get collected telemetry data of segment
     fn get_telemetry_data(&self, detail: TelemetryDetail) -> SegmentTelemetry;
 
     fn fill_query_context(&self, query_context: &mut QueryContext);
 }
 
-#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
-pub struct SegmentManifest {
-    pub segment_version: SeqNumberType,
-    pub file_versions: HashMap<PathBuf, FileVersion>,
-}
-
-#[derive(
-    Copy, Clone, Debug, Eq, PartialEq, Ord, PartialOrd, serde::Serialize, serde::Deserialize,
-)]
-#[serde(untagged)]
-pub enum FileVersion {
-    Version(SeqNumberType),
-    Unversioned,
-}
-
-impl FileVersion {
-    pub fn is_unversioned(self) -> bool {
-        self == Self::Unversioned
-    }
-
-    pub fn or_segment_version(self, segment_version: SeqNumberType) -> SeqNumberType {
-        match self {
-            FileVersion::Version(version) => version,
-            FileVersion::Unversioned => segment_version,
-        }
-    }
-}
-
-impl From<SeqNumberType> for FileVersion {
-    fn from(version: SeqNumberType) -> Self {
-        Self::Version(version)
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn file_version_serde() {
-        test_file_version_serde(FileVersion::Version(42), "42");
-        test_file_version_serde(FileVersion::Unversioned, "null");
-    }
-
-    fn test_file_version_serde(version: FileVersion, json: &str) {
-        let serialized =
-            serde_json::to_string(&version).expect("failed to serialize FileVersion to JSON");
-
-        assert_eq!(serialized, json);
-
-        let deserialized: FileVersion =
-            serde_json::from_str(json).expect("failed to deserialize FileVersion from JSON");
-
-        assert_eq!(deserialized, version);
-    }
-}
