+++ lib/segment/src/entry/mod.rs
@@ -1,1 +1,2 @@
 pub mod entry_point;
+pub mod partial_snapshot_entry;
